+++ feature/notetaking/build.gradle.kts
@@ -21,9 +21,10 @@ dependencies {
     implementation(projects.core.domain)
     implementation(projects.core.data)
     implementation(projects.core.designsystem)
-    implementation("io.legere:pdfiumandroid:1.0.19")
-    implementation("com.itextpdf:itext7-core:7.1.16")
-    implementation("io.legere:pdfium-android-kt-arrow:1.0.19")
+//    implementation("io.legere:pdfiumandroid:1.0.19")
+ //   implementation("com.itextpdf:itext7-core:7.1.16")
+//    implementation("io.legere:pdfium-android-kt-arrow:1.0.19")
+    implementation(libs.pdfbox)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
     implementation(libs.androidx.activity)
