+++ feature/notetaking/src/main/AndroidManifest.xml
@@ -3,8 +3,8 @@
 
     <application>
         <activity
-            android:name=".PdfViewerActivity"
-            android:exported="true" />
+            android:name=".PDFBoxActivity"
+            android:exported="false" />
         <activity
             android:name=".SampleFilePickerViewModel"
             android:exported="false" />
