+++ feature/notetaking/src/main/java/com/iguana/notetaking/SampleFilePickerActivity.kt
@@ -11,7 +11,7 @@ import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.activity.viewModels
-import com.iguana.notetaking.PdfViewerActivity
+//import com.iguana.notetaking.PdfViewerActivity
 
 
 class SampleFilePickerActivity : AppCompatActivity() {
@@ -41,8 +41,8 @@ class SampleFilePickerActivity : AppCompatActivity() {
                     // PDF 메타데이터 가져오기 (iText 외부 라이브러리 사용시 )
                     viewModel.getAdvancedPdfMetadata(this, uri)
 
-                    // 선택된 파일을 PdfViewerActivity로 전달
-                    val intent = Intent(this, PdfViewerActivity::class.java)
+                    // 선택된 파일을 PDFBox Activity 로 전달
+                    val intent = Intent(this, PDFBoxActivity::class.java)
                     intent.putExtra("PDF_URI", uri.toString())  // Uri를 전달
                     startActivity(intent)
                 } else {
