+++ feature/notetaking/src/main/java/com/iguana/notetaking/SampleFilePickerViewModel.kt
@@ -13,7 +13,7 @@ import android.widget.Toast
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.itextpdf.kernel.pdf.PdfReader
+//import com.itextpdf.kernel.pdf.PdfReader
 import java.io.File
 import java.io.FileOutputStream
 
@@ -72,22 +72,22 @@ class SampleFilePickerViewModel : ViewModel() {
     }
     fun getAdvancedPdfMetadata(context: Context, uri: Uri) {
         val inputStream = context.contentResolver.openInputStream(uri)
-        val pdfDoc = com.itextpdf.kernel.pdf.PdfDocument(PdfReader(inputStream))
-
-        // PDF 메타데이터 가져오기
-        val title = pdfDoc.documentInfo.title
-        val author = pdfDoc.documentInfo.author
-        val subject = pdfDoc.documentInfo.subject
-        val keywords = pdfDoc.documentInfo.keywords
-        val creationDate = pdfDoc.documentInfo.getMoreInfo("CreationDate")
-        val modificationDate = pdfDoc.documentInfo.getMoreInfo("ModDate")
-        val producer = pdfDoc.documentInfo.producer
-
-        // 로그에 출력
-        Log.d("PDF_INFO", "(iText 사용해서 받아온 메타데이터) 제목: $title, 저자: $author, 주제: $subject, 키워드: $keywords")
-        Log.d("PDF_INFO", "(iText 사용해서 받아온 메타데이터) 생성일: $creationDate, 수정일: $modificationDate, 생성 소프트웨어: $producer")
-
-        pdfDoc.close()
+//        val pdfDoc = com.itextpdf.kernel.pdf.PdfDocument(PdfReader(inputStream))
+//
+//        // PDF 메타데이터 가져오기
+//        val title = pdfDoc.documentInfo.title
+//        val author = pdfDoc.documentInfo.author
+//        val subject = pdfDoc.documentInfo.subject
+//        val keywords = pdfDoc.documentInfo.keywords
+//        val creationDate = pdfDoc.documentInfo.getMoreInfo("CreationDate")
+//        val modificationDate = pdfDoc.documentInfo.getMoreInfo("ModDate")
+//        val producer = pdfDoc.documentInfo.producer
+//
+//        // 로그에 출력
+//        Log.d("PDF_INFO", "(iText 사용해서 받아온 메타데이터) 제목: $title, 저자: $author, 주제: $subject, 키워드: $keywords")
+//        Log.d("PDF_INFO", "(iText 사용해서 받아온 메타데이터) 생성일: $creationDate, 수정일: $modificationDate, 생성 소프트웨어: $producer")
+//
+//        pdfDoc.close()
     }
 
     fun getFileName(context: Context, uri: Uri): String {
