+++ gradle/libs.versions.toml
@@ -13,6 +13,8 @@ gradleDependencyGraphGeneratorPlugin = "0.8.0"
 room = "2.5.2"
 hilt = "2.46.1"
 hiltNavigationCompose = "1.0.0"
+pdfbox = "2.0.27.0"
+jp2 = "1.0.3"
 
 okhttp = "4.11.0"
 retrofit = "2.9.0"
@@ -134,6 +136,8 @@ material = { group = "com.google.android.material", name = "material", version.r
 androidx-activity = { group = "androidx.activity", name = "activity", version.ref = "activity" }
 androidx-constraintlayout = { group = "androidx.constraintlayout", name = "constraintlayout", version.ref = "constraintlayout" }
 appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+pdfbox = { group = "com.tom-roush", name = "pdfbox-android", version.ref = "pdfbox"}
+jp2-android = { group = "com.gemalto.jp2", name = "jp2-android", version.ref = "jp2"}
 
 [bundles]
